requirements.txt
streamlit
pandas
openpyxl
reportlab
python-dateutil
pywin32